// src/proto/xudrpc_pb.d.ts
    export type AsObject = {
    }
// ...
export class NodeIdentifier extends jspb.Message { 
    getNodePubKey(): string;
    setNodePubKey(value: string): NodeIdentifier;

    getAlias(): string;
    setAlias(value: string): NodeIdentifier;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): NodeIdentifier.AsObject;
    static toObject(includeInstance: boolean, msg: NodeIdentifier): NodeIdentifier.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: NodeIdentifier, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): NodeIdentifier;
    static deserializeBinaryFromReader(message: NodeIdentifier, reader: jspb.BinaryReader): NodeIdentifier;
}

export namespace NodeIdentifier {
    export type AsObject = {
        nodePubKey: string,
        alias: string,
    }
}
// ...
export class Order extends jspb.Message { 
    getPrice(): number;
    setPrice(value: number): Order;

    getQuantity(): number;
    setQuantity(value: number): Order;

    getPairId(): string;
    setPairId(value: string): Order;

    getId(): string;
    setId(value: string): Order;


    hasNodeIdentifier(): boolean;
    clearNodeIdentifier(): void;
    getNodeIdentifier(): NodeIdentifier | undefined;
    setNodeIdentifier(value?: NodeIdentifier): Order;

    getLocalId(): string;
    setLocalId(value: string): Order;

    getCreatedAt(): number;
    setCreatedAt(value: number): Order;

    getSide(): OrderSide;
    setSide(value: OrderSide): Order;

    getIsOwnOrder(): boolean;
    setIsOwnOrder(value: boolean): Order;

    getHold(): number;
    setHold(value: number): Order;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Order.AsObject;
    static toObject(includeInstance: boolean, msg: Order): Order.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: Order, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Order;
    static deserializeBinaryFromReader(message: Order, reader: jspb.BinaryReader): Order;
}

export namespace Order {
    export type AsObject = {
        price: number,
        quantity: number,
        pairId: string,
        id: string,
        nodeIdentifier?: NodeIdentifier.AsObject,
        localId: string,
        createdAt: number,
        side: OrderSide,
        isOwnOrder: boolean,
        hold: number,
    }
}
// ...
export class Orders extends jspb.Message { 
    clearBuyOrdersList(): void;
    getBuyOrdersList(): Array<Order>;
    setBuyOrdersList(value: Array<Order>): Orders;
    addBuyOrders(value?: Order, index?: number): Order;

    clearSellOrdersList(): void;
    getSellOrdersList(): Array<Order>;
    setSellOrdersList(value: Array<Order>): Orders;
    addSellOrders(value?: Order, index?: number): Order;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): Orders.AsObject;
    static toObject(includeInstance: boolean, msg: Orders): Orders.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: Orders, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): Orders;
    static deserializeBinaryFromReader(message: Orders, reader: jspb.BinaryReader): Orders;
}

export namespace Orders {
    export type AsObject = {
        buyOrdersList: Array<Order.AsObject>,
        sellOrdersList: Array<Order.AsObject>,
    }
}
// ...
export enum OrderSide {
    BUY = 0,
    SELL = 1,
    BOTH = 2,
}